+++ LightCrafter/src/LCR_Command_Packetizer.h
@@ -12,15 +12,15 @@ using namespace std;
 class Command_Packetizer
 {
 private:
-	void InitilizeCommandBuffer(uint8* command, uint8 packetType, uint16 commandId, uint8 flags, uint16 payLoadLength);
-	void LoadPayLoadInBuffer (uint8 * command, uint8* payLoad, uint16 payLoadLength);
-	void CalculateCheckSum(uint8* command, uint16 size);
+	void InitilizeCommandBuffer(uint8* command, uint8 packetType, uint16 commandId, uint8 flags, long payLoadLength);
+	void LoadPayLoadInBuffer (uint8 * command, uint8* payLoad, long payLoadLength);
+	void CalculateCheckSum(uint8* command, long size);
 
 public:
 	Command_Packetizer(void);
 	~Command_Packetizer(void);
 
-	uint8* CreateCommand(uint8 packetType, uint16 commandId, uint8 flags, uint16 payLoadLength, uint8* payLoad);
+	uint8* CreateCommand(uint8 packetType, uint16 commandId, uint8 flags, long payLoadLength, uint8* payLoad);
 
 };
